refactor(plugin): migrate to new api package
